Move config file class out of stack
