+++ app/models/runtime/stack.rb
@@ -1,4 +1,5 @@
 require 'models/helpers/process_types'
+require 'models/helpers/config_file'
 
 module VCAP::CloudController
   class Stack < Sequel::Model
@@ -95,41 +96,6 @@ module VCAP::CloudController
         Steno.logger('cc.stack').warn('stack.populate.collision', hash) if stack.modified?
       else
         create(hash.slice('name', 'description', 'build_rootfs_image', 'run_rootfs_image'))
-      end
-    end
-
-    class ConfigFile
-      def initialize(file_path)
-        @hash = YAMLConfig.safe_load_file(file_path).tap do |h|
-          Schema.validate(h)
-        end
-      end
-
-      def stacks
-        @hash['stacks']
-      end
-
-      def deprecated_stacks
-        @hash['deprecated_stacks']
-      end
-
-      def default
-        @hash['default']
-      end
-
-      Schema = Membrane::SchemaParser.parse do
-        {
-          'default' => String,
-          'stacks' => [{
-            'name' => String,
-            'description' => String,
-            optional('build_rootfs_image') => String,
-            optional('run_rootfs_image') => String
-          }],
-          optional('deprecated_stacks') => [
-            String
-          ]
-        }
       end
     end
   end
