+++ lib/tasks/stack_check.rake
@@ -7,6 +7,7 @@ namespace :stacks do
     next unless db.table_exists?(:buildpack_lifecycle_data)
 
     RakeConfig.config.load_db_encryption_key
+    require 'models/helpers/config_file'
     require 'cloud_controller/check_stacks'
     VCAP::CloudController::CheckStacks.new(RakeConfig.config, db).validate_stacks
   end
